Small refactoring of calibration folder.
