+++ transformersurgeon/calibration/__init__.py
@@ -1,6 +1,8 @@
 """Calibration framework package."""
 
+# Public entrypoint used by the manager to run the full calibration pass.
 from .backbone import run_compression_calibration
+# Public list of summary names currently supported by the framework.
 from .summaries import SUPPORTED_SUMMARIES
 
 __all__ = [
