+++ transformersurgeon/calibration/raw_data/registry.py
@@ -7,13 +7,16 @@ from .weight_grad import WeightGradCollector
 
 
 RAW_DATA_REGISTRY = {
+    # Activation stream used by covariance-like summaries.
     "activation": ActivationCollector,
+    # Weight-gradient stream used by gradient-aware pruning methods.
     "weight_grad": WeightGradCollector,
 }
 SUPPORTED_RAW_DATA = tuple(RAW_DATA_REGISTRY.keys())
 
 
 def instantiate_raw_collector(raw_name: str, *, offload_to_cpu: bool):
+    # Factory used by the backbone to instantiate one collector per raw stream type.
     try:
         collector_type = RAW_DATA_REGISTRY[raw_name]
     except KeyError as exc:
