+++ transformersurgeon/calibration/summaries/base.py
@@ -12,11 +12,14 @@ import torch
 class CalibrationSummary(ABC):
     """Abstract summary contract used by the calibration backbone."""
 
+    # Registry key used by compressors in needs_calibration().
     name: str
+    # Raw streams required before one summary update can happen.
     required_raw_data: Tuple[str, ...]
 
     def initialize_store(self, calibration_store: dict) -> None:
         """Prepare storage keys owned by this summary."""
+        # Reset only this summary namespace, keep other summaries untouched.
         calibration_store.pop(self.name, None)
 
     @abstractmethod
@@ -25,6 +28,7 @@ class CalibrationSummary(ABC):
 
     def update_runtime(self, runtime, calibration_store: dict, raw_data: Mapping[str, torch.Tensor]) -> None:
         """Runtime-aware update hook for summaries that need internal state."""
+        # Default implementation ignores runtime state and performs a direct overwrite.
         self.update_from_raw(calibration_store, raw_data)
 
     def finalize_store(self, calibration_store: dict) -> None:
@@ -41,12 +45,15 @@ class SummaryRuntime:
     state: Dict[str, int] = field(default_factory=dict)
 
     def reset_batch(self) -> None:
+        # Clear any partial payload from previous batch before new raw emissions arrive.
         self.pending_raw.clear()
 
     def on_raw(self, calibration_store: dict, raw_name: str, raw_value: torch.Tensor) -> None:
+        # Ignore raw streams unrelated to this summary.
         if raw_name not in self.summary.required_raw_data:
             return
         self.pending_raw[raw_name] = raw_value
+        # Update only when we have a complete payload for this summary.
         if all(name in self.pending_raw for name in self.summary.required_raw_data):
             self.summary.update_runtime(self, calibration_store, self.pending_raw)
             self.num_updates += 1
@@ -54,6 +61,7 @@ class SummaryRuntime:
 
 
 def unique_summaries(summary_names: Iterable[str]) -> Tuple[str, ...]:
+    # Order-preserving de-duplication for deterministic calibration behavior.
     unique = []
     for summary_name in summary_names:
         if summary_name not in unique:
